requirements.txt
streamlit>=1.28.0
streamlit-chat>=0.1.1
pytesseract>=0.3.10
Pillow>=10.0.0
# ollama-client  # or groq (if using Groq API)
easyocr>=1.7.0
serpapi>=0.1.5
newspaper3k>=0.2.8
tinydb>=4.8.0
sqlite-utils>=3.35
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.28.0
nomic>=2.0.0
duckduckgo-search>=4.1.0
gcsa>=2.0.0
markdown>=3.5.0
google-generativeai>=0.3.0
openai>=1.0.0
PyPDF2>=3.0.0 